+++ transforms/remToFloat.js
@@ -1,4 +1,5 @@
 module.exports = {
+  name: 'size/remToFloat',
   type: 'value',
   matcher: (token) => token.attributes.category === 'size',
   transformer: (token) => {
